[Tolk] Breaking change: separate `address` and `any_address`
Before: 'address' meant internal/external/none.
Now: 'address' is internal only, 'any_address' is internal/external/none.

`address?` (nullable) is slice/null on a stack, but serialized as internal/none
(267 or 2 bits).

With new TVM 12 instructions (LDSTDADDR and others), it's now possible to load
and check whether an address is internal without extra gas cost. That's the reason
why it was not done before.