[Tolk] Better deal with exhaustive `match`, enums, and malformed stack
Before, `match (enum)` covering all enum members, was just compiled as IF ELSE,
the last ELSE was a fallback for invalid (malformed) enum members also. In practice,
this can occur only if a programmer intentionally disables compiler checks
or corrupts the stack via asm functions.
Anyway, I decided to add an extra IF, so that if a stack contains an invalid value,
excno 5 was thrown.